Added firmware encryption tool

+++ build.sh
@@ -13,6 +13,16 @@ usage() {
     exit 1
 }
 
+# check the required tools
+check_tools "awk zip app_version.sh app_model.sh ack2_swu_encrypt.py python3"
+
+# set the custom encrypt tool
+ENCRYPT_TOOL=$(which "ack2_swu_encrypt.py")
+if [ -z "$ENCRYPT_TOOL" ]; then
+    # if not installed use the local copy
+    ENCRYPT_TOOL="TOOLS/ack2_swu_encrypt.py"
+fi
+
 # selected fw file
 selected_firmware_file=""
 
@@ -198,10 +208,29 @@ fi
 
 # Process the output file if set
 if [ -n "$build_output" ]; then
+
+    # try to find out the app version (like app_ver="3.1.0")
+    def_target="$ROOTFS_DIR/app/app"
+    app_ver=$("$app_version_tool" "$def_target")
+    if [ $? != 0 ]; then
+        echo -e "${RED}ERROR: Cannot find the app version ${NC}"
+        exit 4
+    fi
+
+    # try to find out the model
+    app_model=$("$app_model_tool" "$def_target")
+    if [ $? != 0 ]; then
+        echo -e "${RED}ERROR: Cannot find the app model ${NC}"
+        exit 5
+    fi
+
+    rm -f "$project_root/update.bin"
     rm -f "$project_root/update.zip"
     zip -r "$project_root/update.zip" update
-    /bin/cp -f "$project_root/update.zip" "$build_output"
+    $ENCRYPT_TOOL -i "$project_root/update.zip" -o "$project_root/update.bin" -m "$app_model" -v "$app_ver"
+    /bin/cp -f "$project_root/update.bin" "$build_output"
     rm -f "$project_root/update.zip"
+    rm -f "$project_root/update.bin"
 fi
 
 echo
